Refactor: Quick Add as child component
