+++ pages/home.tsx
@@ -1,33 +1,17 @@
 import type { NextPage } from 'next'
-import { KeyboardEvent, useState } from 'react'
-import { useQuery, useMutation } from '@apollo/client'
-import {
-  CreateTaskMutation,
-  CreateTaskDocument,
-  GetTasksQuery,
-  GetTasksDocument,
-} from '@/graphql/types/client'
+import { useState } from 'react'
+import { useQuery } from '@apollo/client'
+import { GetTasksQuery, GetTasksDocument } from '@/graphql/types/client'
 import { useSession } from 'next-auth/react'
 import Header from '@/components/header'
 import Board from '@/components/board'
 import { TaskCards } from '@/components/task_card'
 import { css } from '@emotion/react'
+import QuickAdd from '@/components/quick_add'
 
 const boards = css`
   display: flex;
   min-height: 50rem;
-`
-
-const addTask = css`
-  color: inherit;
-  box-sizing: border-box;
-  margin-bottom: 1rem;
-  padding: 12px 16px;
-  border-color: transparent;
-  border-radius: 4px;
-  background-color: #424242;
-  width: 100%;
-  font-size: inherit;
 `
 
 const Home: NextPage = () => {
@@ -38,26 +22,7 @@ const Home: NextPage = () => {
   })
   //TODO: rename inputValue
   const [inputValue, setInputValue] = useState('')
-  const [createTask] = useMutation<CreateTaskMutation>(CreateTaskDocument, {
-    onCompleted() {
-      refetch()
-      setInputValue('')
-    },
-  })
   const [selectedTaskId, setSelectedTaskId] = useState('')
-
-  const handleKeyDown = (event: KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === 'Enter' && inputValue) {
-      createTask({
-        variables: {
-          task: {
-            title: inputValue,
-            userId: session?.user?.id,
-          },
-        },
-      })
-    }
-  }
 
   const openTaskDetail = (taskId: string | undefined) => {
     if (taskId) {
@@ -70,16 +35,12 @@ const Home: NextPage = () => {
       <Header />
       <div css={boards}>
         <Board>
-          <div>
-            <input
-              type="text"
-              value={inputValue}
-              placeholder="+ add new task"
-              onChange={(e) => setInputValue(e.target.value)}
-              onKeyDown={handleKeyDown}
-              css={addTask}
-            ></input>
-          </div>
+          <QuickAdd
+            inputValue={inputValue}
+            setInputValue={setInputValue}
+            userId={session?.user?.id}
+            refetch={refetch}
+          />
           <TaskCards
             data={data}
             refetch={refetch}
